incident_resolution_backend/requirements.txt
flask
python-dotenv
psycopg2-binary
sentence_transformers